+++ CHANGELOG.md
@@ -20,6 +20,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 
 --- -->
+## [0.4.6] - 2026-03-06
+
+### Added
+
+- Add createInventoryItem function to handle item creation in Airtable
+- Add saving state management to QuickAddForm and AddInventoryItemForm
+- Implement throttled fetch for Airtable API to manage rate limits and handle errors
+- Add saveError state management to handle inline error messages in forms
+- Display saving status message in MainContainer when saving an item to Airtable
+- Integrate createInventoryItem function to handle item submissions in MainContainer
+
+### Changed
+
+- Refactor AddInventoryItemForm and QuickAddForm to integrate createInventoryItem function for handling submissions
+- Refactor AddInventoryItemForm and QuickAddForm to simplify item addition and remove unused state management
+
+---
 
 ## [0.4.5] - 2026-03-06